refactor: extract RTK rewrite logic into rtk-rewrite.ts
bash-collapse.ts was doing two unrelated things: collapsing multiline
commands for display, and rewriting commands through rtk before execution.
Move the RTK concern (detectRtk, rewriteWithRtk, isRtkPassthrough,
rtkSpawnHook, rewritePreparedBashCommand, the extension default export)
into a dedicated rtk-rewrite.ts / rtk-rewrite.test.ts.

bash-collapse.ts now contains only collapseCommand.

Consumer updates:
- src/cli.ts: import rtkRewriteExtension from rtk-rewrite
- src/extensions/tool-rendering.ts: import getBashCommandForDisplay from rtk-rewrite

No behaviour changes.
